Don't carryover org-journal items
As long as the old TODOs stay on agenda and can be navigated to through
it, I don't need to carry over. My entries are defined in terms of time,
and it's ok when it's done within one file (read day), but when you
begin to mix multiple days' times, things become messy.

Turns out you can use directory for org-agenda-files, and it will
include all files inside the directory.