Promote 50/72 formatting for commit messages
Summary max length may be not enough, especially with Conventional
Commits, but it's not a hard limit - it just reminds you to consider
making your message shorter.

fill-column, on the other hand, is enforced via auto-fill-mode, and
obviously I have no clue where and why Emacs got 70 (that is used by
default).