Configure task's category in the agenda view
This requires to set CATEGORY property for TODO entries, otherwise it
will be just the file name of the entry (and we only have one file right
now, so it doesn't add a lot of information).

What I do is to set it for each project to its name (the verb may be
omitted - we are restricted in space here after all), and for the Tasks
entry to "Tasks". This way, and with the default property inheritance
for CATEGORY in the agenda context, for each TODO I see the project it
belongs to (or "Tasks" if it doesn't belong to any).