Calculate line-numbers area width size upfront
When using display-line-numbers-type 'relative, chances are that your
current line needs more digits for representation than the the maximum
relative count of the lines your display can accommodate. This may
result in the current line always being indented relative to the other
lines, and it's not the best UX when moving through lines. This commit
sets line-numbers area width to the maximum number of digits requiring
to represent a line number in the current buffer.