Prevent straight from loading the built-in transient
This partially reverts d1f1059. I don't need it, as I found what was
causing the magit error ("transient-setup: Symbol’s function definition
is void: transient-prefix-object"). From a certain version, it obviously
depends on transient-prefix-object (that's a relatively new function
compared to v0.4.3 that is the built-in version). But then, even though
straight pulls and build the latest version, I believe it first
activates the built-in, and afterwards "(require 'transient')" returns
immediately, because the package is already loaded. I tried unloading
and loading it again, but it failed with some slot error. So I tried
to (use-package transient) as early as possible, because from straight
docs I saw that if you explicitly don't do :type built-in, it will
download the latest available version. And it worked. And putting it
within any of the other functions didn't.